pyproject.toml
[tool.poetry]
name = "sandbox-cdf"
version = "0.1.0"
description = "Playground for Computational Fluid Dynamics"
authors = ["Emil"]

[tool.poetry.dependencies]
python = "^3.9"
numpy = {file = "../3PP/wheels/numpy-1.19.2+mkl-cp39-cp39-win_amd64.whl"}
scipy = {file = "../3PP/wheels/scipy-1.5.3-cp39-cp39-win_amd64.whl"}
findiff = "^0.8.5"
matplotlib = "^3.3.2"
jupyterlab = "^2.2.9"
jupyter-tabnine = "^1.1.0"
Pillow = "^8.0.1"

[tool.poetry.dev-dependencies]
black = "^20.8b1"
pytest = "^6.1.1"
pylint = "^2.6.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
